View Models, Models, States and Protocols refactor - removal of bloat code, isoltating Models into smaller pieces, using class for models and using short hand codes
