+++ voltixApp/VoltixApp/Model/BitcoinTransactionMempool.swift
@@ -15,25 +15,14 @@ class BitcoinTransactionMempool: Codable {
         set { _userAddress = newValue }
     }
     
-    init(txid: String, version: Int, locktime: Int, vin: [Input], vout: [Output], fee: Int, status: TransactionStatus, userAddress: String) {
-        self.txid = txid
-        self.version = version
-        self.locktime = locktime
-        self.vin = vin
-        self.vout = vout
-        self.fee = fee
-        self.status = status
-        self.userAddress = userAddress
-    }
-    
     var isSent: Bool {
-        return vin.contains { input in
+        vin.contains { input in
             input.prevout?.scriptpubkey_address == userAddress
         }
     }
     
     var isReceived: Bool {
-        return vout.contains { output in
+        vout.contains { output in
             output.scriptpubkey_address == userAddress
         }
     }
@@ -62,37 +51,14 @@ class BitcoinTransactionMempool: Codable {
         return totalSentToOthers
     }
     
-    class Input: Codable {
-        let txid: String
-        let vout: Int
-        let prevout: PreviousOutput?
-        let sequence: UInt32
-        let scriptsig: String?
-        let scriptsig_asm: String?
-        let witness: [String]?
-        let is_coinbase: Bool?
-        
-        class PreviousOutput: Codable {
-            let scriptpubkey: String
-            let scriptpubkey_asm: String
-            let scriptpubkey_type: String
-            let scriptpubkey_address: String?
-            let value: Int
-        }
-    }
-    
-    class Output: Codable {
-        let scriptpubkey: String
-        let scriptpubkey_asm: String
-        let scriptpubkey_type: String
-        let scriptpubkey_address: String?
-        let value: Int
-    }
-    
-    class TransactionStatus: Codable {
-        let confirmed: Bool
-        let block_height: Int?
-        let block_hash: String?
-        let block_time: Int?
+    init(txid: String, version: Int, locktime: Int, vin: [Input], vout: [Output], fee: Int, status: TransactionStatus, userAddress: String) {
+        self.txid = txid
+        self.version = version
+        self.locktime = locktime
+        self.vin = vin
+        self.vout = vout
+        self.fee = fee
+        self.status = status
+        self.userAddress = userAddress
     }
 }
