+++ voltixApp/VoltixApp/Model/KeyShare.swift
@@ -1,0 +1,13 @@
+//
+//  KeyShare.swift
+//  VoltixApp
+//
+//  Created by Amol Kumar on 2024-03-04.
+//
+
+import Foundation
+
+struct KeyShare: Codable {
+    let pubkey: String
+    let keyshare: String
+}
