+++ voltixApp/VoltixApp/Model/Output.swift
@@ -1,0 +1,16 @@
+//
+//  Output.swift
+//  VoltixApp
+//
+//  Created by Amol Kumar on 2024-03-04.
+//
+
+import Foundation
+
+class Output: Codable {
+    let scriptpubkey: String
+    let scriptpubkey_asm: String
+    let scriptpubkey_type: String
+    let scriptpubkey_address: String?
+    let value: Int
+}
