+++ voltixApp/VoltixApp/Protocols/DecodableDefaultSource.swift
@@ -3,12 +3,9 @@
 //  VoltixApp
 //
 
-
 import Foundation
 
 protocol DecodableDefaultSource {
     associatedtype Value: Decodable
     static var defaultValue: Value { get }
 }
-
-enum DecodableDefault {}
