+++ voltixApp/VoltixApp/States/AppNavigationState.swift
@@ -7,7 +7,6 @@ import Foundation
 import SwiftUI
 
 enum CurrentScreen: Hashable {
-    
     // Onboarding
     case welcome // Welcome screen
     case startScreen  // New or Import wallet
@@ -27,12 +26,12 @@ enum CurrentScreen: Hashable {
     case JoinKeysign
     
     // Normal use (typically launches here if wallet imported/generated already)
-    case vaultAssets(SendTransaction)  // Main landing page for normal use. Lists ETH, BTC, ... assets.
+    case vaultAssets(SendTransaction) // Main landing page for normal use. Lists ETH, BTC, ... assets.
     case menu  // Add/Export/Forget vaults
     
     // Send
     case sendInputDetails(SendTransaction)
-    case sendVerifyScreen(SendTransaction)     // 2nd device goes to here automatically on receiving a p2p keysign msg
+    case sendVerifyScreen(SendTransaction) // 2nd device goes to here automatically on receiving a p2p keysign msg
     
     // Swap
     case swapInputDetails
@@ -41,7 +40,6 @@ enum CurrentScreen: Hashable {
     case swapVerifyScreen
     case swapDone
     
-
     case listVaultAssetView
     
     // transactions
