+++ voltixApp/VoltixApp/States/DecodableDefault.swift
@@ -1,0 +1,10 @@
+//
+//  DecodableDefault.swift
+//  VoltixApp
+//
+//  Created by Amol Kumar on 2024-03-04.
+//
+
+import Foundation
+
+enum DecodableDefault {}
