+++ voltixApp/VoltixApp/States/SupportedAssets.swift
@@ -10,7 +10,7 @@ enum AssetType: CaseIterable {
     case bitcoin
     case ethereum
     case thorchain
-
+    
     var ticker: String {
         switch self {
         case .bitcoin:
@@ -21,7 +21,7 @@ enum AssetType: CaseIterable {
             "RUNE"
         }
     }
-
+    
     var chainName: String {
         switch self {
         case .bitcoin:
@@ -31,19 +31,5 @@ enum AssetType: CaseIterable {
         case .thorchain:
             "THORChain"
         }
-    }
-}
-
-struct Asset : Codable,Hashable {
-    let ticker: String
-    let chainName: String
-    let image: String
-    let contractAddress: String?
-    
-    init(ticker: String, chainName: String, image: String, contractAddress: String?) {
-        self.ticker = ticker
-        self.chainName = chainName
-        self.image = image
-        self.contractAddress = contractAddress
     }
 }
